마이페이지 수정 / 페이지형식으로 변경 / 프로필 네비게이트 경로 수정
