+++ src/Router.js
@@ -4,6 +4,7 @@ import RegisterLocation from "./pages/RegisterLocation";
 import Register from "./pages/Register";
 import Login from "./pages/Login";
 import Main from "./pages/Main";
+import MyPage from "./pages/MyPage";
 import Add from "./pages/Add";
 import Detail from "./pages/Detail";
 import NotFound from "./pages/NotFound";
@@ -20,6 +21,7 @@ function Router() {
         <Route path="/register" element={<Register />} />
         <Route path="/login" element={<Login />} />
         <Route path="/add" element={<Add />} />
+        <Route path="/mypage" element={<MyPage />} />
         <Route path="/modify/:postid" element={<Modify />} />
         <Route path="/detail/:postid" element={<Detail />} />
         <Route path="/profile/location" element={<ProfileLocation />} />
